Create module module
