+++ pyqir/src/lib.rs
@@ -7,6 +7,7 @@ mod context;
 mod evaluator;
 mod generator;
 mod instructions;
+mod module;
 mod python;
 mod simple;
 mod types;
